Minor changes
Vertically focussed videos were overlapping with the Ninja's
volumebar/sprite; so made minor changes to try to stop that.
Looking at the css, it is clear that logically grouping HTML elements
using divs, while making the Javascript simpler, it makes the css
unbearable. If time permits, it could be overhauled, but that is
doubtful.